Manually cleans "other problems".
Removes duplicate dependent tiers.

***********
Special attention to the file SUP14WM.cha: this commit adds a line where a
person speaks "1, 2, 3". The reason is that there is a line

"%eng: one, two, three"

just afterwards. Instead of removing the line, I decided to create an utterance
(which is supposed to be there), spoken by an unidentified speaker.
***********

 Changes to be committed:
	modified:   ALI31TF.XXS
	modified:   ALI78TF.XXS
	modified:   MAE34TF.XXS
	modified:   MAE41All.NAC
	modified:   MAE51WM.NAC
	modified:   SUP11WM.NAC
	modified:   SUP14WM.NAC
	modified:   SUP33TF.XXS
	modified:   SUP43TF.XXS
	modified:   SUP51WM.NAC
	modified:   SUP52WM.NAC
	modified:   SUP75TF.XXS
	modified:   SUP84All.NAC