Add scripts for cleaning inuktitut data.
`prepare_for_chatter.sh` runs all other scripts in the required order, and
afterwards inserts a '\r' in the end of the resulting files (so that Chatter
doesn't complain). Three folders will be created. The folder that must be passed
to chatter is the one name "output-timfix".

`run_acqdiv` runs the "acqdiv clean" and then runs `check` in its output,
creating a folder called "check" with the output of `check` of each file.

`run_timfix.sh` runs `timfix.py` in each file and creates a folder called
"output-timfix". It also creates a folder called "output", which is a copy of
the output of "acqdiv clean".

 Changes to be committed:
	new file:   prepare_for_chatter.sh
	new file:   run_acqdiv.sh
	new file:   run_timfix.sh
	new file:   timfix.py
	deleted:    ../timfix.py