Insert metada for SUP62.
While creating the scripts, it became clear that the metadata for SUP62 was
missing. It had been lost after the merging of the files from all same
session.

The metadata for SUP62 that was inserted in this commit was "recovered" from the
first metadata file Robert had sent me in the very beginning of the Inuktitut
cleaning process.

 Changes to be committed:
	modified:   ../sessions.csv